RF-12: addressing comments
