+++ src/app/navigation/auth.tsx
@@ -5,8 +5,8 @@ import { LoginScreen } from 'src/app/screens/auth/login';
 import { BackArrowIcon } from 'src/assets/svg/navigation-icons';
 
 import { Header } from '../components/header';
-import { CompleteDataScreen } from '../screens/auth/signup/completeData';
-import { CreateAccountScreen } from '../screens/auth/signup/createAccount';
+import { CompleteDataScreen } from '../screens/auth/signup/complete-data';
+import { CreateAccountScreen } from '../screens/auth/signup/create-account';
 import { RootStack } from './root-stack';
 
 export const getAuthenticatedGroup = () => {
