+++ src/app/screens/auth/signup/create-account/index.tsx
@@ -9,40 +9,36 @@ import {
 } from 'react-native';
 import { Box, Button, ButtonText, Text } from '@gluestack-ui/themed';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { useNavigation } from '@react-navigation/native';
+import { StackScreenProps } from '@react-navigation/stack';
 
-import { ControlledInput } from 'src/app/components/inputs';
+import { ControlledInput, PasswordInput } from 'src/app/components/inputs';
+import { useAppDispatch } from 'src/store';
+import { startCreateFirebaseUser } from 'src/store/auth/thunks';
+import { RootStackParamList } from 'src/types/navigation';
 import { commonStyles } from 'src/utils/styles';
 
-import { validationSchema } from './form-config';
+import { FormData, validationSchema } from './form-config';
 
-type FormData = {
-  firstName: string;
-  lastName: string;
-  birthDate: string;
-  gender: string;
-};
-
-export const CompleteDataScreen = () => {
-  const { navigate } = useNavigation<any>();
+export const CreateAccountScreen = ({
+  navigation,
+}: StackScreenProps<RootStackParamList, 'CreateAccount'>) => {
   const { t } = useTranslation();
+  const dispatch = useAppDispatch();
 
   const { control, handleSubmit } = useForm<FormData>({
     defaultValues: {
-      firstName: '',
-      lastName: '',
-      birthDate: '',
-      gender: undefined,
+      email: '',
+      password: '',
+      repeatPassword: '',
     },
     resolver: yupResolver(validationSchema),
   });
 
   const onValidSubmit: SubmitHandler<FormData> = async data => {
     try {
-      // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      const { firstName, lastName, birthDate, gender } = data;
-      // TODO: dispatch create userInfo
-      navigate('Login');
+      const { email, password } = data;
+      await dispatch(startCreateFirebaseUser({ email, password }));
+      navigation.navigate('CompleteData');
     } catch (error: any) {
       Alert.alert(t('screens:signUp:error'), error.message);
     }
@@ -54,55 +50,44 @@ export const CompleteDataScreen = () => {
         behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
         style={commonStyles.keyboardAvoidingView}
       >
-        <Box padding={20} flex={1} justifyContent="center" mt="$6" mb="$8">
-          <Text size="2xl" textAlign="center" mb="$4">
-            {t('screens:completeData.heading1')}
+        <Box padding={20} flex={1} justifyContent="center" mb="$12">
+          <Text size="2xl" textAlign="center" mb="$12">
+            {t('screens:signUp.heading1')}
           </Text>
           <ControlledInput
             controller={{
               control,
-              name: 'firstName',
+              name: 'email',
             }}
             formControlProps={{
               mb: '$4',
             }}
           />
-          <ControlledInput
+          <PasswordInput
             controller={{
               control,
-              name: 'lastName',
+              name: 'password',
             }}
             formControlProps={{
               mb: '$4',
             }}
           />
-          {/* TODO: datepicker implementation */}
-          <ControlledInput
+          <PasswordInput
             controller={{
               control,
-              name: 'birthDate',
+              name: 'repeatPassword',
             }}
             formControlProps={{
               mb: '$4',
             }}
           />
-          {/* TODO: implement a select input */}
-          <ControlledInput
-            controller={{
-              control,
-              name: 'gender',
-            }}
-            formControlProps={{
-              mb: '$4',
-            }}
-          />
-
           <Button
             onPress={handleSubmit(onValidSubmit)}
-            mt="$4"
+            mt="$8"
+            mb="$4"
             bgColor="$lime600"
           >
-            <ButtonText>{t('common:button.confirm')}</ButtonText>
+            <ButtonText>{t('common:button.continue')}</ButtonText>
           </Button>
         </Box>
       </KeyboardAvoidingView>
